Add simple gui to federate and setup test federate
Add a simple gui to the federate that allows the user to join, resign, and
execute the gameloop, and shows status messages.

Setup the skf test federate.
This is what should be modified going forward.
It currently has the orbit class added in and calls the propogate method
once per update cycle.
The orbit class now needs to be changed to fit our moon/satellite scenario
instead of the one used in the elliptical orbit guide.
http://www.castor2.ca/04_Propagation/index.html